Code modularization and some optimizations

+++ include/ant.h
@@ -26,7 +26,7 @@ class Ant
 		//int getChoice() const { return _choice; }
 		//enum State { REST, SEARCH_NEW_NESTBOX, FIND_NESTBOX, BACK_TO_HOME, DANCE };
 		//State getState() const { return _state; }
-		void environmentAnalysis(int viewFrequency, vector<int> &pheromoneMatrix);
+		void environmentAnalysis(int viewFrequency, vector<uint8_t> &pheromoneMatrix);
 		void putPheromone(int* pheromoneMatrix);
 		void move(int l);
 		void draw();
